Use discrete stats and rolls

+++ src/main/scala/swords/Creature.scala
@@ -2,7 +2,7 @@ package swords
 
 final case class Creature(
   position: V2[Int],
-  hitPoints: Double,
-  attack: Double,
-  defense: Double
+  hitPoints: Int,
+  attack: Int,
+  defense: Int
 )
