+++ src/main/scala/swords/GameEvent.scala
@@ -5,7 +5,7 @@ sealed trait GameEvent
 final case class CreatureAttacked(
   attackerName: String,
   defenderName: String,
-  damage: Option[Double]
+  damage: Option[Int]
 ) extends GameEvent
 
 final case class CreatureDied(name: String) extends GameEvent
